Move highlights from fold to inlay randomized tests
